crypto_app/crypto_app_backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
charset-normalizer==2.0.12
click==8.2.1
colorama==0.4.6
Django==3.2.10
fastapi==0.116.1
flake8==7.3.0
Flask==2.0.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.26.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
